+++ source/cl/source/binary/source/spirv.cpp
@@ -28,7 +28,7 @@ namespace binary {
 // TODO: add a proper mechanism for extending spirv-ll and reporting extension
 // support. We don't actually support the generic storage class extension on
 // all core targets. See CA-3067.
-const std::array<const std::string, 10> supported_extensions = {
+const std::array<const std::string, 11> supported_extensions = {
     {
         "SPV_KHR_no_integer_wrap_decoration",
         "SPV_INTEL_kernel_attributes",
@@ -40,6 +40,7 @@ const std::array<const std::string, 10> supported_extensions = {
         "SPV_INTEL_arbitrary_precision_integers",
         "SPV_INTEL_optnone",
         "SPV_INTEL_memory_access_aliasing",
+        "SPV_INTEL_subgroups",
     },
 };
 
@@ -51,7 +52,7 @@ cargo::expected<compiler::spirv::DeviceInfo, cargo::result> getSPIRVDeviceInfo(
   auto &spvCapabilities = spvDeviceInfo.capabilities;
 
   // A set of capabilities shared between the OpenCL profiles we support.
-  static std::array<spv::Capability, 14> sharedCapabilities = {
+  static std::array<spv::Capability, 15> sharedCapabilities = {
       spv::CapabilityAddresses,
       spv::CapabilityFloat16Buffer,
       spv::CapabilityGroups,
@@ -66,6 +67,7 @@ cargo::expected<compiler::spirv::DeviceInfo, cargo::result> getSPIRVDeviceInfo(
       spv::CapabilityArbitraryPrecisionIntegersINTEL,
       spv::CapabilityOptNoneINTEL,
       spv::CapabilityMemoryAccessAliasingINTEL,
+      spv::CapabilitySubgroupShuffleINTEL,
   };
 
   if (profile == "FULL_PROFILE") {